backend/requirements.txt
alembic==1.7.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
coverage==6.3.1
cryptography==36.0.1
defusedxml==0.7.1
Flask==2.0.3
Flask-Cors==3.0.10
flask-ssl==0.0.0.1rc1
Flask-SSLify==0.1.5
genbadge==1.0.6
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mypy==0.931
mypy-extensions==0.4.3
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0